Updated VOCASY dll and implementations.
Vocasy library now uses UnityEngine.dll and contains basic general
components that can be used by all implemetations and unity specific
objects. Vocasy library now contains tests for its classes. Removed test
scripts from implementations alongside all the objects/components that
have been inglobed by the library

+++ SteamVersion/Assets/VoiceChat/Scripts/Voice Chat/Transports/SteamLobbyAudioTransport.cs
@@ -10,7 +10,7 @@ public class SteamLobbyAudioTransport : MonoBehaviour, IAudioTransportLayer
 
     public bool IsPacketAvailable { get { uint pack; return SteamNetworking.IsP2PPacketAvailable(out pack); } }
 
-    public uint MaxPacketLength { get { return (uint)temp1024.MaxCapacity - FirstPacketByteAvailable; } }
+    public int MaxPacketLength { get { return temp1024.MaxCapacity - FirstPacketByteAvailable; } }
 
     [SerializeField]
     private SteamPlayer voiceUserPrefab;
